Update routes add pagination (#85)
Updated routes, removed old methods
