+++ sdk/transaction_model.go
@@ -14,9 +14,9 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/google/flatbuffers/go"
-	"github.com/proximax-storage/go-xpx-crypto"
-	"github.com/proximax-storage/go-xpx-utils"
+	flatbuffers "github.com/google/flatbuffers/go"
+	crypto "github.com/proximax-storage/go-xpx-crypto"
+	utils "github.com/proximax-storage/go-xpx-utils"
 
 	"github.com/proximax-storage/go-xpx-chain-sdk/transactions"
 )
@@ -109,45 +109,28 @@ func (tx *AbstractTransaction) buildVectors(builder *flatbuffers.Builder, v uint
 	transactions.TransactionBufferAddDeadline(builder, dV)
 }
 
-type abstractTransactionDTO struct {
-	Type      EntityType              `json:"type"`
-	Version   int64                   `json:"version"`
-	MaxFee    *uint64DTO              `json:"maxFee"`
-	Deadline  *blockchainTimestampDTO `json:"deadline"`
-	Signature string                  `json:"signature"`
-	Signer    string                  `json:"signer"`
+type Pagination struct {
+	TotalEntries uint64
+	PageNumber   uint64
+	PageSize     uint64
+	TotalPages   uint64
 }
 
-func (dto *abstractTransactionDTO) toStruct(tInfo *TransactionInfo) (*AbstractTransaction, error) {
-	nt := ExtractNetworkType(dto.Version)
+type TransactionsPage struct {
+	Transactions []Transaction
+	Pagination   Pagination
+}
 
-	tv := EntityVersion(ExtractVersion(dto.Version))
-
-	pa, err := NewAccountFromPublicKey(dto.Signer, nt)
-	if err != nil {
-		return nil, err
-	}
-
-	var d *Deadline
-	if dto.Deadline != nil {
-		d = NewDeadlineFromBlockchainTimestamp(dto.Deadline.toStruct())
-	}
-
-	var f Amount
-	if dto.MaxFee != nil {
-		f = dto.MaxFee.toStruct()
-	}
-
-	return &AbstractTransaction{
-		*tInfo,
-		nt,
-		d,
-		dto.Type,
-		tv,
-		f,
-		dto.Signature,
-		pa,
-	}, nil
+type TransactionsPageOptions struct {
+	Height           uint   `url:"Height,omitempty"`
+	FromHeight       uint64 `url:"FromHeight,omitempty"`
+	ToHeight         uint64 `url:"ToHeight,omitempty"`
+	Address          string `url:"Address,omitempty"`
+	SignerPublicKey  string `url:"SignerPublicKey,omitempty"`
+	RecipientAddress string `url:"RecipientAddress,omitempty"`
+	Type             []uint `url:"Type,omitempty"`
+	Embedded         bool   `url:"Embedded,omitempty"`
+	PaginationOrderingOptions
 }
 
 type TransactionInfo struct {
@@ -182,49 +165,6 @@ func (ti *TransactionInfo) String() string {
 		ti.UniqueAggregateHash,
 		ti.AggregateId,
 	)
-}
-
-type transactionInfoDTO struct {
-	Height              uint64DTO `json:"height"`
-	Index               uint32    `json:"index"`
-	Id                  string    `json:"id"`
-	TransactionHash     hashDto   `json:"hash"`
-	MerkleComponentHash hashDto   `json:"merkleComponentHash"`
-	AggregateHash       hashDto   `json:"aggregateHash,omitempty"`
-	UniqueAggregateHash hashDto   `json:"uniqueAggregateHash,omitempty"`
-	AggregateId         string    `json:"aggregateId,omitempty"`
-}
-
-func (dto *transactionInfoDTO) toStruct() (*TransactionInfo, error) {
-	transactionHash, err := dto.TransactionHash.Hash()
-	if err != nil {
-		return nil, err
-	}
-	merkleComponentHash, err := dto.MerkleComponentHash.Hash()
-	if err != nil {
-		return nil, err
-	}
-	aggregateHash, err := dto.AggregateHash.Hash()
-	if err != nil {
-		return nil, err
-	}
-	uniqueAggregateHash, err := dto.UniqueAggregateHash.Hash()
-	if err != nil {
-		return nil, err
-	}
-
-	ref := TransactionInfo{
-		dto.Height.toStruct(),
-		dto.Index,
-		dto.Id,
-		transactionHash,
-		merkleComponentHash,
-		aggregateHash,
-		uniqueAggregateHash,
-		dto.AggregateId,
-	}
-
-	return &ref, nil
 }
 
 type AccountPropertiesAddressModification struct {
@@ -332,59 +272,6 @@ func (tx *AccountPropertiesAddressTransaction) Size() int {
 	return AccountPropertyAddressHeader + (AccountPropertiesAddressModificationSize * len(tx.Modifications))
 }
 
-type accountPropertiesAddressModificationDTO struct {
-	ModificationType PropertyModificationType `json:"type"`
-	Address          string                   `json:"value"`
-}
-
-func (dto *accountPropertiesAddressModificationDTO) toStruct() (*AccountPropertiesAddressModification, error) {
-	a, err := NewAddressFromBase32(dto.Address)
-	if err != nil {
-		return nil, err
-	}
-
-	return &AccountPropertiesAddressModification{
-		dto.ModificationType,
-		a,
-	}, nil
-}
-
-type accountPropertiesAddressTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		PropertyType  PropertyType                               `json:"propertyType"`
-		Modifications []*accountPropertiesAddressModificationDTO `json:"modifications"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *accountPropertiesAddressTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	ms := make([]*AccountPropertiesAddressModification, len(dto.Tx.Modifications))
-	for i, m := range dto.Tx.Modifications {
-		ms[i], err = m.toStruct()
-
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return &AccountPropertiesAddressTransaction{
-		*atx,
-		dto.Tx.PropertyType,
-		ms,
-	}, nil
-}
-
 type AccountPropertiesMosaicModification struct {
 	ModificationType PropertyModificationType
 	AssetId          AssetId
@@ -485,59 +372,6 @@ func (tx *AccountPropertiesMosaicTransaction) Bytes() ([]byte, error) {
 
 func (tx *AccountPropertiesMosaicTransaction) Size() int {
 	return AccountPropertyMosaicHeader + (AccountPropertiesMosaicModificationSize * len(tx.Modifications))
-}
-
-type accountPropertiesMosaicModificationDTO struct {
-	ModificationType PropertyModificationType `json:"type"`
-	AssetId          assetIdDTO               `json:"value"`
-}
-
-func (dto *accountPropertiesMosaicModificationDTO) toStruct() (*AccountPropertiesMosaicModification, error) {
-	assetId, err := dto.AssetId.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	return &AccountPropertiesMosaicModification{
-		dto.ModificationType,
-		assetId,
-	}, nil
-}
-
-type accountPropertiesMosaicTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		PropertyType  PropertyType                              `json:"propertyType"`
-		Modifications []*accountPropertiesMosaicModificationDTO `json:"modifications"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *accountPropertiesMosaicTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	ms := make([]*AccountPropertiesMosaicModification, len(dto.Tx.Modifications))
-	for i, m := range dto.Tx.Modifications {
-		ms[i], err = m.toStruct()
-
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return &AccountPropertiesMosaicTransaction{
-		*atx,
-		dto.Tx.PropertyType,
-		ms,
-	}, nil
 }
 
 type AccountPropertiesEntityTypeModification struct {
@@ -642,54 +476,6 @@ func (tx *AccountPropertiesEntityTypeTransaction) Size() int {
 	return AccountPropertyEntityTypeHeader + (AccountPropertiesEntityModificationSize * len(tx.Modifications))
 }
 
-type accountPropertiesEntityTypeModificationDTO struct {
-	ModificationType PropertyModificationType `json:"type"`
-	EntityType       EntityType               `json:"value"`
-}
-
-func (dto *accountPropertiesEntityTypeModificationDTO) toStruct() (*AccountPropertiesEntityTypeModification, error) {
-	return &AccountPropertiesEntityTypeModification{
-		dto.ModificationType,
-		dto.EntityType,
-	}, nil
-}
-
-type accountPropertiesEntityTypeTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		PropertyType  PropertyType                                  `json:"propertyType"`
-		Modifications []*accountPropertiesEntityTypeModificationDTO `json:"modifications"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *accountPropertiesEntityTypeTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	ms := make([]*AccountPropertiesEntityTypeModification, len(dto.Tx.Modifications))
-	for i, m := range dto.Tx.Modifications {
-		ms[i], err = m.toStruct()
-
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return &AccountPropertiesEntityTypeTransaction{
-		*atx,
-		dto.Tx.PropertyType,
-		ms,
-	}, nil
-}
-
 type AliasTransaction struct {
 	AbstractTransaction
 	ActionType  AliasActionType
@@ -735,30 +521,6 @@ func (tx *AliasTransaction) Size() int {
 
 func (tx *AliasTransaction) GetAbstractTransaction() *AbstractTransaction {
 	return &tx.AbstractTransaction
-}
-
-type aliasTransactionDTO struct {
-	abstractTransactionDTO
-	NamespaceId namespaceIdDTO  `json:"namespaceId"`
-	ActionType  AliasActionType `json:"aliasAction"`
-}
-
-func (dto *aliasTransactionDTO) toStruct(tInfo *TransactionInfo) (*AliasTransaction, error) {
-	atx, err := dto.abstractTransactionDTO.toStruct(tInfo)
-	if err != nil {
-		return nil, err
-	}
-
-	namespaceId, err := dto.NamespaceId.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	return &AliasTransaction{
-		*atx,
-		dto.ActionType,
-		namespaceId,
-	}, nil
 }
 
 type AddressAliasTransaction struct {
@@ -820,36 +582,6 @@ func (tx *AddressAliasTransaction) Size() int {
 	return tx.AliasTransaction.Size() + AddressSize
 }
 
-type addressAliasTransactionDTO struct {
-	Tx struct {
-		aliasTransactionDTO
-		Address string `json:"address"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *addressAliasTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.aliasTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	a, err := NewAddressFromBase32(dto.Tx.Address)
-	if err != nil {
-		return nil, err
-	}
-
-	return &AddressAliasTransaction{
-		*atx,
-		a,
-	}, nil
-}
-
 type MosaicAliasTransaction struct {
 	AliasTransaction
 	MosaicId *MosaicId
@@ -904,36 +636,6 @@ func (tx *MosaicAliasTransaction) Bytes() ([]byte, error) {
 
 func (tx *MosaicAliasTransaction) Size() int {
 	return tx.AliasTransaction.Size() + MosaicIdSize
-}
-
-type mosaicAliasTransactionDTO struct {
-	Tx struct {
-		aliasTransactionDTO
-		MosaicId *mosaicIdDTO `json:"mosaicId"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *mosaicAliasTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.aliasTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	mosaicId, err := dto.Tx.MosaicId.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	return &MosaicAliasTransaction{
-		*atx,
-		mosaicId,
-	}, nil
 }
 
 type AccountLinkTransaction struct {
@@ -1003,38 +705,6 @@ func (tx *AccountLinkTransaction) Bytes() ([]byte, error) {
 
 func (tx *AccountLinkTransaction) Size() int {
 	return AccountLinkTransactionSize
-}
-
-type accountLinkTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		RemoteAccountKey string            `json:"remoteAccountKey"`
-		Action           AccountLinkAction `json:"action"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *accountLinkTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	acc, err := NewAccountFromPublicKey(dto.Tx.RemoteAccountKey, atx.NetworkType)
-	if err != nil {
-		return nil, err
-	}
-
-	return &AccountLinkTransaction{
-		*atx,
-		acc,
-		dto.Tx.Action,
-	}, nil
 }
 
 type NetworkConfigTransaction struct {
@@ -1125,51 +795,6 @@ func (tx *NetworkConfigTransaction) Size() int {
 	return NetworkConfigHeaderSize + len(tx.NetworkConfig.String()) + len(tx.SupportedEntities.String())
 }
 
-type networkConfigTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		ApplyHeightDelta        uint64DTO `json:"applyHeightDelta"`
-		NetworkConfig           string    `json:"networkConfig"`
-		SupportedEntityVersions string    `json:"supportedEntityVersions"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *networkConfigTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	applyHeightDelta := dto.Tx.ApplyHeightDelta.toUint64()
-
-	s := NewSupportedEntities()
-
-	err = s.UnmarshalBinary([]byte(dto.Tx.SupportedEntityVersions))
-	if err != nil {
-		return nil, err
-	}
-
-	c := NewNetworkConfig()
-
-	err = c.UnmarshalBinary([]byte(dto.Tx.NetworkConfig))
-	if err != nil {
-		return nil, err
-	}
-
-	return &NetworkConfigTransaction{
-		*atx,
-		Duration(applyHeightDelta),
-		c,
-		s,
-	}, nil
-}
-
 type BlockchainUpgradeTransaction struct {
 	AbstractTransaction
 	UpgradePeriod        Duration
@@ -1232,36 +857,6 @@ func (tx *BlockchainUpgradeTransaction) Bytes() ([]byte, error) {
 
 func (tx *BlockchainUpgradeTransaction) Size() int {
 	return BlockchainUpgradeTransactionSize
-}
-
-type blockchainUpgradeTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		UpgradePeriod        uint64DTO `json:"upgradePeriod"`
-		NewBlockChainVersion uint64DTO `json:"newBlockChainVersion"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *blockchainUpgradeTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	upgradePeriod := dto.Tx.UpgradePeriod.toUint64()
-	newBlockChainVersion := dto.Tx.NewBlockChainVersion.toUint64()
-
-	return &BlockchainUpgradeTransaction{
-		*atx,
-		Duration(upgradePeriod),
-		BlockChainVersion(newBlockChainVersion),
-	}, nil
 }
 
 type AggregateTransaction struct {
@@ -1381,61 +976,6 @@ func (tx *AggregateTransaction) Size() int {
 	return AggregateBondedHeaderSize + sizeOfInnerTransactions
 }
 
-type aggregateTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		Cosignatures      []*aggregateTransactionCosignatureDTO `json:"cosignatures"`
-		InnerTransactions []map[string]interface{}              `json:"transactions"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *aggregateTransactionDTO) toStruct(generationHash *Hash) (Transaction, error) {
-	txsr, err := json.Marshal(dto.Tx.InnerTransactions)
-	if err != nil {
-		return nil, err
-	}
-
-	txs, err := MapTransactions(bytes.NewBuffer(txsr), generationHash)
-	if err != nil {
-		return nil, err
-	}
-
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	as := make([]*AggregateTransactionCosignature, len(dto.Tx.Cosignatures))
-	for i, a := range dto.Tx.Cosignatures {
-		as[i], err = a.toStruct(atx.NetworkType)
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	for _, tx := range txs {
-		iatx := tx.GetAbstractTransaction()
-		iatx.Deadline = atx.Deadline
-		iatx.Signature = atx.Signature
-		iatx.MaxFee = atx.MaxFee
-		iatx.TransactionInfo = atx.TransactionInfo
-	}
-
-	agtx := AggregateTransaction{
-		*atx,
-		txs,
-		as,
-	}
-
-	return &agtx, agtx.UpdateUniqueAggregateHash(generationHash)
-}
-
 type ModifyMetadataTransaction struct {
 	AbstractTransaction
 	MetadataType  MetadataType
@@ -1493,30 +1033,6 @@ func (tx *ModifyMetadataTransaction) GetAbstractTransaction() *AbstractTransacti
 	return &tx.AbstractTransaction
 }
 
-type modifyMetadataTransactionDTO struct {
-	abstractTransactionDTO
-	MetadataType  MetadataType               `json:"metadataType"`
-	Modifications []*metadataModificationDTO `json:"modifications"`
-}
-
-func (dto *modifyMetadataTransactionDTO) toStruct(tInfo *TransactionInfo) (*ModifyMetadataTransaction, error) {
-	atx, err := dto.abstractTransactionDTO.toStruct(tInfo)
-	if err != nil {
-		return nil, err
-	}
-
-	ms, err := metadataDTOArrayToStruct(dto.Modifications, atx.NetworkType)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ModifyMetadataTransaction{
-		*atx,
-		dto.MetadataType,
-		ms,
-	}, nil
-}
-
 type ModifyMetadataAddressTransaction struct {
 	ModifyMetadataTransaction
 	Address *Address
@@ -1572,36 +1088,6 @@ func (tx *ModifyMetadataAddressTransaction) Size() int {
 	return tx.ModifyMetadataTransaction.Size() + AddressSize
 }
 
-type modifyMetadataAddressTransactionDTO struct {
-	Tx struct {
-		modifyMetadataTransactionDTO
-		Address string `json:"metadataId"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *modifyMetadataAddressTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.modifyMetadataTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	a, err := NewAddressFromBase32(dto.Tx.Address)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ModifyMetadataAddressTransaction{
-		*atx,
-		a,
-	}, nil
-}
-
 type ModifyMetadataMosaicTransaction struct {
 	ModifyMetadataTransaction
 	MosaicId *MosaicId
@@ -1654,36 +1140,6 @@ func (tx *ModifyMetadataMosaicTransaction) Size() int {
 	return tx.ModifyMetadataTransaction.Size() + MosaicIdSize
 }
 
-type modifyMetadataMosaicTransactionDTO struct {
-	Tx struct {
-		modifyMetadataTransactionDTO
-		MosaicId *mosaicIdDTO `json:"metadataId"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *modifyMetadataMosaicTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.modifyMetadataTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	mosaicId, err := dto.Tx.MosaicId.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	return &ModifyMetadataMosaicTransaction{
-		*atx,
-		mosaicId,
-	}, nil
-}
-
 type ModifyMetadataNamespaceTransaction struct {
 	ModifyMetadataTransaction
 	NamespaceId *NamespaceId
@@ -1734,36 +1190,6 @@ func (tx *ModifyMetadataNamespaceTransaction) Bytes() ([]byte, error) {
 
 func (tx *ModifyMetadataNamespaceTransaction) Size() int {
 	return tx.ModifyMetadataTransaction.Size() + NamespaceSize
-}
-
-type modifyMetadataNamespaceTransactionDTO struct {
-	Tx struct {
-		modifyMetadataTransactionDTO
-		NamespaceId *namespaceIdDTO `json:"metadataId"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *modifyMetadataNamespaceTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.modifyMetadataTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	namespaceId, err := dto.Tx.NamespaceId.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	return &ModifyMetadataNamespaceTransaction{
-		*atx,
-		namespaceId,
-	}, nil
 }
 
 type MosaicDefinitionTransaction struct {
@@ -1868,45 +1294,6 @@ func (tx *MosaicDefinitionTransaction) Size() int {
 	return MosaicDefinitionTransactionHeaderSize + len(tx.OptionalProperties)*MosaicOptionalPropertySize
 }
 
-type mosaicDefinitionTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		Properties  mosaicPropertiesDTO `json:"properties"`
-		MosaicNonce int64               `json:"mosaicNonce"`
-		MosaicId    *mosaicIdDTO        `json:"mosaicId"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *mosaicDefinitionTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	mosaicId, err := dto.Tx.MosaicId.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	properties, err := dto.Tx.Properties.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	return &MosaicDefinitionTransaction{
-		*atx,
-		properties,
-		uint32(dto.Tx.MosaicNonce),
-		mosaicId,
-	}, nil
-}
-
 type MosaicSupplyChangeTransaction struct {
 	AbstractTransaction
 	MosaicSupplyType
@@ -1981,40 +1368,6 @@ func (tx *MosaicSupplyChangeTransaction) Bytes() ([]byte, error) {
 
 func (tx *MosaicSupplyChangeTransaction) Size() int {
 	return MosaicSupplyChangeTransactionSize
-}
-
-type mosaicSupplyChangeTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		MosaicSupplyType `json:"direction"`
-		AssetId          *assetIdDTO `json:"mosaicId"`
-		Delta            uint64DTO   `json:"delta"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *mosaicSupplyChangeTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	assetId, err := dto.Tx.AssetId.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	return &MosaicSupplyChangeTransaction{
-		*atx,
-		dto.Tx.MosaicSupplyType,
-		assetId,
-		dto.Tx.Delta.toStruct(),
-	}, nil
 }
 
 type TransferTransaction struct {
@@ -2154,56 +1507,6 @@ func (tx *TransferTransaction) MessageSize() int {
 	return len(tx.Message.Payload()) + 1
 }
 
-type transferTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		Message messageDTO   `json:"message"`
-		Mosaics []*mosaicDTO `json:"mosaics"`
-		Address string       `json:"recipient"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *transferTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	mosaics := make([]*Mosaic, len(dto.Tx.Mosaics))
-
-	for i, mosaic := range dto.Tx.Mosaics {
-		msc, err := mosaic.toStruct()
-		if err != nil {
-			return nil, err
-		}
-
-		mosaics[i] = msc
-	}
-
-	a, err := NewAddressFromBase32(dto.Tx.Address)
-	if err != nil {
-		return nil, err
-	}
-
-	m, err := dto.Tx.Message.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	return &TransferTransaction{
-		*atx,
-		m,
-		mosaics,
-		a,
-	}, nil
-}
-
 type ModifyMultisigAccountTransaction struct {
 	AbstractTransaction
 	MinApprovalDelta int8
@@ -2279,40 +1582,6 @@ func (tx *ModifyMultisigAccountTransaction) Bytes() ([]byte, error) {
 
 func (tx *ModifyMultisigAccountTransaction) Size() int {
 	return ModifyMultisigHeaderSize + ((KeySize + 1 /* MultisigModificationType size */) * len(tx.Modifications))
-}
-
-type modifyMultisigAccountTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		MinApprovalDelta int8                                  `json:"minApprovalDelta"`
-		MinRemovalDelta  int8                                  `json:"minRemovalDelta"`
-		Modifications    []*multisigCosignatoryModificationDTO `json:"modifications"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *modifyMultisigAccountTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	ms, err := multisigCosignatoryDTOArrayToStruct(dto.Tx.Modifications, atx.NetworkType)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ModifyMultisigAccountTransaction{
-		*atx,
-		dto.Tx.MinApprovalDelta,
-		dto.Tx.MinRemovalDelta,
-		ms,
-	}, nil
 }
 
 type ModifyContractTransaction struct {
@@ -2429,59 +1698,6 @@ func (tx *ModifyContractTransaction) Bytes() ([]byte, error) {
 
 func (tx *ModifyContractTransaction) Size() int {
 	return ModifyContractHeaderSize + ((KeySize + 1 /* ContractModificationType size */) * (len(tx.Customers) + len(tx.Executors) + len(tx.Verifiers)))
-}
-
-type modifyContractTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		DurationDelta uint64DTO                             `json:"durationDelta"`
-		Hash          hashDto                               `json:"hash"`
-		Customers     []*multisigCosignatoryModificationDTO `json:"customers"`
-		Executors     []*multisigCosignatoryModificationDTO `json:"executors"`
-		Verifiers     []*multisigCosignatoryModificationDTO `json:"verifiers"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *modifyContractTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	customers, err := multisigCosignatoryDTOArrayToStruct(dto.Tx.Customers, atx.NetworkType)
-	if err != nil {
-		return nil, err
-	}
-
-	executors, err := multisigCosignatoryDTOArrayToStruct(dto.Tx.Executors, atx.NetworkType)
-	if err != nil {
-		return nil, err
-	}
-
-	verifiers, err := multisigCosignatoryDTOArrayToStruct(dto.Tx.Verifiers, atx.NetworkType)
-	if err != nil {
-		return nil, err
-	}
-
-	hash, err := dto.Tx.Hash.Hash()
-	if err != nil {
-		return nil, err
-	}
-
-	return &ModifyContractTransaction{
-		*atx,
-		dto.Tx.DurationDelta.toStruct(),
-		hash,
-		customers,
-		executors,
-		verifiers,
-	}, nil
 }
 
 type RegisterNamespaceTransaction struct {
@@ -2603,56 +1819,6 @@ func (tx *RegisterNamespaceTransaction) Size() int {
 	return RegisterNamespaceHeaderSize + len(tx.NamspaceName)
 }
 
-type registerNamespaceTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		Id            namespaceIdDTO `json:"namespaceId"`
-		NamespaceType `json:"namespaceType"`
-		NamspaceName  string    `json:"name"`
-		Duration      uint64DTO `json:"duration"`
-		ParentId      namespaceIdDTO
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *registerNamespaceTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	d := Duration(0)
-	var n *NamespaceId = nil
-
-	if dto.Tx.NamespaceType == Root {
-		d = dto.Tx.Duration.toStruct()
-	} else {
-		n, err = dto.Tx.ParentId.toStruct()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	nsId, err := dto.Tx.Id.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	return &RegisterNamespaceTransaction{
-		*atx,
-		nsId,
-		dto.Tx.NamespaceType,
-		dto.Tx.NamspaceName,
-		d,
-		n,
-	}, nil
-}
-
 type LockFundsTransaction struct {
 	AbstractTransaction
 	*Mosaic
@@ -2734,51 +1900,6 @@ func (tx *LockFundsTransaction) Bytes() ([]byte, error) {
 
 func (tx *LockFundsTransaction) Size() int {
 	return LockSize
-}
-
-type lockFundsTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		AssetId  assetIdDTO `json:"mosaicId"`
-		Amount   uint64DTO  `json:"amount"`
-		Duration uint64DTO  `json:"duration"`
-		Hash     hashDto    `json:"hash"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *lockFundsTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	assetId, err := dto.Tx.AssetId.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	mosaic, err := NewMosaic(assetId, dto.Tx.Amount.toStruct())
-	if err != nil {
-		return nil, err
-	}
-
-	hash, err := dto.Tx.Hash.Hash()
-	if err != nil {
-		return nil, err
-	}
-
-	return &LockFundsTransaction{
-		*atx,
-		mosaic,
-		dto.Tx.Duration.toStruct(),
-		&SignedTransaction{AggregateBonded, "", hash},
-	}, nil
 }
 
 type SecretLockTransaction struct {
@@ -2877,59 +1998,6 @@ func (tx *SecretLockTransaction) Size() int {
 	return SecretLockSize
 }
 
-type secretLockTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		AssetId   *assetIdDTO `json:"mosaicId"`
-		Amount    *uint64DTO  `json:"amount"`
-		HashType  HashType    `json:"hashAlgorithm"`
-		Duration  uint64DTO   `json:"duration"`
-		Secret    string      `json:"secret"`
-		Recipient string      `json:"recipient"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *secretLockTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	a, err := NewAddressFromBase32(dto.Tx.Recipient)
-	if err != nil {
-		return nil, err
-	}
-
-	assetId, err := dto.Tx.AssetId.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	mosaic, err := NewMosaic(assetId, dto.Tx.Amount.toStruct())
-	if err != nil {
-		return nil, err
-	}
-
-	secret, err := NewSecretFromHexString(dto.Tx.Secret, dto.Tx.HashType)
-	if err != nil {
-		return nil, err
-	}
-
-	return &SecretLockTransaction{
-		*atx,
-		mosaic,
-		dto.Tx.Duration.toStruct(),
-		secret,
-		a,
-	}, nil
-}
-
 type SecretProofTransaction struct {
 	AbstractTransaction
 	HashType
@@ -3017,45 +2085,6 @@ func (tx *SecretProofTransaction) Size() int {
 	return SecretProofHeaderSize + tx.Proof.Size()
 }
 
-type secretProofTransactionDTO struct {
-	Tx struct {
-		abstractTransactionDTO
-		HashType  `json:"hashAlgorithm"`
-		Proof     string `json:"proof"`
-		Recipient string `json:"recipient"`
-	} `json:"transaction"`
-	TDto transactionInfoDTO `json:"meta"`
-}
-
-func (dto *secretProofTransactionDTO) toStruct(*Hash) (Transaction, error) {
-	info, err := dto.TDto.toStruct()
-	if err != nil {
-		return nil, err
-	}
-
-	atx, err := dto.Tx.abstractTransactionDTO.toStruct(info)
-	if err != nil {
-		return nil, err
-	}
-
-	proof, err := NewProofFromHexString(dto.Tx.Proof)
-	if err != nil {
-		return nil, err
-	}
-
-	a, err := NewAddressFromBase32(dto.Tx.Recipient)
-	if err != nil {
-		return nil, err
-	}
-
-	return &SecretProofTransaction{
-		*atx,
-		dto.Tx.HashType,
-		proof,
-		a,
-	}, nil
-}
-
 type CosignatureTransaction struct {
 	TransactionToCosign *AggregateTransaction
 }
@@ -3127,22 +2156,6 @@ type AggregateTransactionCosignature struct {
 	Signer    *PublicAccount
 }
 
-type aggregateTransactionCosignatureDTO struct {
-	Signature string `json:"signature"`
-	Signer    string
-}
-
-func (dto *aggregateTransactionCosignatureDTO) toStruct(networkType NetworkType) (*AggregateTransactionCosignature, error) {
-	acc, err := NewAccountFromPublicKey(dto.Signer, networkType)
-	if err != nil {
-		return nil, err
-	}
-	return &AggregateTransactionCosignature{
-		dto.Signature,
-		acc,
-	}, nil
-}
-
 func (agt *AggregateTransactionCosignature) String() string {
 	return fmt.Sprintf(
 		`
@@ -3170,23 +2183,6 @@ func (m *MultisigCosignatoryModification) String() string {
 	)
 }
 
-type multisigCosignatoryModificationDTO struct {
-	Type          MultisigCosignatoryModificationType `json:"type"`
-	PublicAccount string                              `json:"cosignatoryPublicKey"`
-}
-
-func (dto *multisigCosignatoryModificationDTO) toStruct(networkType NetworkType) (*MultisigCosignatoryModification, error) {
-	acc, err := NewAccountFromPublicKey(dto.PublicAccount, networkType)
-	if err != nil {
-		return nil, err
-	}
-
-	return &MultisigCosignatoryModification{
-		dto.Type,
-		acc,
-	}, nil
-}
-
 type MetadataModification struct {
 	Type  MetadataModificationType
 	Key   string
@@ -3210,23 +2206,9 @@ func (m *MetadataModification) String() string {
 	)
 }
 
-type metadataModificationDTO struct {
-	Type  MetadataModificationType `json:"modificationType"`
-	Key   string                   `json:"key"`
-	Value string                   `json:"value"`
-}
-
-func (dto *metadataModificationDTO) toStruct(networkType NetworkType) (*MetadataModification, error) {
-	return &MetadataModification{
-		dto.Type,
-		dto.Key,
-		dto.Value,
-	}, nil
-}
-
 type TransactionStatus struct {
 	Deadline *Deadline
-	Group    string
+	Group    TransactionGroup
 	Status   string
 	Hash     *Hash
 	Height   Height
@@ -3247,37 +2229,6 @@ func (ts *TransactionStatus) String() string {
 		ts.Deadline,
 		ts.Height,
 	)
-}
-
-type transactionStatusDTO struct {
-	Group    string                 `json:"group"`
-	Status   string                 `json:"status"`
-	Hash     hashDto                `json:"hash"`
-	Deadline blockchainTimestampDTO `json:"deadline"`
-	Height   uint64DTO              `json:"height"`
-}
-
-func (dto *transactionStatusDTO) toStruct() (*TransactionStatus, error) {
-	hash, err := dto.Hash.Hash()
-	if err != nil {
-		return nil, err
-
-	}
-	return &TransactionStatus{
-		NewDeadlineFromBlockchainTimestamp(dto.Deadline.toStruct()),
-		dto.Group,
-		dto.Status,
-		hash,
-		dto.Height.toStruct(),
-	}, nil
-}
-
-type TransactionIdsDTO struct {
-	Ids []string `json:"transactionIds"`
-}
-
-type TransactionHashesDTO struct {
-	Hashes []string `json:"hashes"`
 }
 
 const (
